move files to reflect the their depenecies throughout the application
