+++ src/showCodeCoverageCommand/decorations/codeCovered.ts
@@ -1,5 +1,5 @@
 import { DecorationOptions, Range, TextDocument, TextEditor } from 'vscode';
-import { Coverage, RSpecResults } from '../types';
+import { Coverage, RSpecResults } from '../../types';
 
 const fileNames = (editors: TextEditor[]) =>
   editors
